Issue/1

+++ lindenmayer-system/lindenmayer-string.h
@@ -38,7 +38,7 @@ public:
 		}
 
 		bool AtEnd() {
-			return m_pos >= m_parent.m_symbols.size();
+			return (unsigned int)m_pos >= m_parent.m_symbols.size();
 		}
 
 		LindenmayerSymbol& operator*() {
